requirements-gravity-api.txt
# Gravity NIL Intelligence API (gravity_api/) — install atop base requirements.
# Scraping and torch-based scoring run in separate repositories; keep this stack light.
fastapi>=0.115.0
uvicorn[standard]>=0.31.0
asyncpg>=0.30.0
anthropic>=0.39.0
numpy>=1.26.0
pandas>=2.2.0
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.9.0
PyJWT>=2.8.0
